Fixed some bugs about Camera and Type Manager, added a first test about damage of entities

+++ src/Components/Battle/LifeBarComponent.h
@@ -6,6 +6,6 @@
 
 struct LifeBarComponent
 {
-	unsigned int health;
-	unsigned int maxHealth;
+	float health;
+	float maxHealth;
 };
